requirements.txt
Django 4.1.4

django-bootstrap-modal-forms 2.2.0

django-widget-tweaks 1.4.12